use size_t consistently

+++ src/main.cpp
@@ -3,7 +3,7 @@
 
 int main(int, char** argv)
 {
-    const auto exponent = std::atoi(argv[1]);
+    const auto exponent = static_cast<std::size_t>(std::atoi(argv[1]));
     std::cout << "p="  << exponent << " -> " << std::boolalpha << is_mersenne_prime(exponent) << std::endl;
     return 0;
 }
